refactor: update configs,ignore and index files
