+++ .babelrc
@@ -1,6 +1,7 @@
 {
     "presets": [
         ["@babel/preset-env", { "targets": { "node": "current" } }],
+        "@babel/preset-typescript",
         "@babel/react"
     ],
     "plugins": ["@babel/transform-runtime"]
