+++ .eslintignore
@@ -1,0 +1,5 @@
+node_module/
+dist/
+webpack.config.web.js
+webpack.config.server.js
+tsconfig.json
