+++ src/web/App.test.tsx
@@ -1,13 +1,9 @@
-import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { configureStore } from '@reduxjs/toolkit';
 import { Provider } from 'react-redux';
 
 import App from './App';
-import { rootReducer } from './store/rootReducer';
-
-const store = configureStore({ reducer: rootReducer });
+import { store } from './state';
 
 describe('App', () => {
     it('renders', () => {
