+++ src/web/main.styles.tsx
@@ -1,18 +1,20 @@
 import React from 'react';
+import { renderToStaticMarkup } from 'react-dom/server';
 import { createGlobalStyle } from 'styled-components';
 
+import Background from './components/Background/Background';
 import UniSansHeavyCAPSEOT from './assets/UniSansHeavyCAPS.eot';
 import UniSansHeavyCAPSWOFF2 from './assets/UniSansHeavyCAPS.woff2';
 import UniSansHeavyCAPSEWOFF from './assets/UniSansHeavyCAPS.woff';
 import UniSansHeavyCAPSETTF from './assets/UniSansHeavyCAPS.ttf';
 
-import { renderToStaticMarkup } from 'react-dom/server';
-import { Background } from './web/components/Background';
 const svgBackground = encodeURIComponent(
     renderToStaticMarkup(
         <Background gradientStart={'#37003c'} gradientEnd={'#963cff'} />
     )
 );
+
+/* eslint-disable import/prefer-default-export */
 
 export const GlobalStyle = createGlobalStyle`
